requirements.txt
streamlit 
rembg 
pillow 
io